Moving the base task to the project root
