+++ pokemongo_bot/cell_workers/follow_spiral.py
@@ -5,7 +5,7 @@ import math
 
 from pokemongo_bot.cell_workers.utils import distance, format_dist
 from pokemongo_bot.step_walker import StepWalker
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 
 class FollowSpiral(BaseTask):
     def initialize(self):
