+++ pokemongo_bot/cell_workers/handle_soft_ban.py
@@ -3,7 +3,7 @@ from random import randint
 from pgoapi.utilities import f2i
 
 from pokemongo_bot.constants import Constants
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 from pokemongo_bot.cell_workers import MoveToFort
 from pokemongo_bot.cell_workers.utils import distance
 from pokemongo_bot.worker_result import WorkerResult
