+++ pokemongo_bot/cell_workers/nickname_pokemon.py
@@ -1,5 +1,5 @@
 from pokemongo_bot.human_behaviour import sleep
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 
 class NicknamePokemon(BaseTask):
     def initialize(self):
