+++ pokemongo_bot/cell_workers/pokemon_catch_worker.py
@@ -3,7 +3,7 @@
 import time
 from pokemongo_bot.human_behaviour import (normalized_reticle_size, sleep,
                                            spin_modifier)
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 
 class PokemonCatchWorker(BaseTask):
     BAG_FULL = 'bag_full'
